Long should work now
git-svn-id: svn://hinux.hin.no/home/DT2006/murray/svn/gh@224 c6f41995-d01f-4a11-924a-98aa25776321

+++ trunk/grouphug/modules/WordCount.java
@@ -52,13 +52,13 @@ public class WordCount implements GrouphugModule {
                 long existingLines = 0;
                 bot.sendMessage("Starting words extraction", false);
                 try {
-                    existingWords = ((Integer)values[1]);
+                    existingWords = ((Long)values[1]);
                 } catch(ClassCastException ex) {
                     bot.sendMessage("Caught WORDS CCE: "+ex, false);
                 }
                 bot.sendMessage("Starting lines extraction", false);
                 try {
-                    existingLines = ((Integer)values[2]);
+                    existingLines = ((Long)values[2]);
                 } catch(ClassCastException ex) {
                     bot.sendMessage("Caught LINES CCE: "+ex, false);
                 }
